Added initialization for tree
Added sex.
Changed some methods to be more efficient.
I think the structure should be a arraylist of arraylists with connections stored with the people, to make traversing up, down and across the tree the easiest. We should focus on the actual probabilities first, before we worry about the names and stuff.